+++ packages/react/src/components/FxEditor/index.tsx
@@ -134,23 +134,19 @@ const FxEditor: React.FC = () => {
   );
   const insertSelectedFormula = useCallback(
     (formulaName: string) => {
-      if (/^=[a-zA-Z]+$/.test(refs.fxInput.current!.innerText)) {
-        const ht = `<span dir="auto" class="luckysheet-formula-text-color">=</span><span dir="auto" class="luckysheet-formula-text-func">${formulaName}</span><span dir="auto" class="luckysheet-formula-text-lpar">(</span>`;
-        refs.fxInput.current!.innerHTML = ht;
-        const cellEditor = document.getElementById(
-          "luckysheet-rich-text-editor"
-        );
-        if (cellEditor) {
-          cellEditor.innerHTML = ht;
-        }
-
-        moveCursorToEnd(refs.fxInput.current!);
-        setContext((draftCtx) => {
-          draftCtx.functionCandidates = [];
-          draftCtx.defaultCandidates = [];
-          draftCtx.functionHint = formulaName;
-        });
+      const ht = `<span dir="auto" class="luckysheet-formula-text-color">=</span><span dir="auto" class="luckysheet-formula-text-func">${formulaName}</span><span dir="auto" class="luckysheet-formula-text-lpar">(</span>`;
+      refs.fxInput.current!.innerHTML = ht;
+      const cellEditor = document.getElementById("luckysheet-rich-text-editor");
+      if (cellEditor) {
+        cellEditor.innerHTML = ht;
       }
+
+      moveCursorToEnd(refs.fxInput.current!);
+      setContext((draftCtx) => {
+        draftCtx.functionCandidates = [];
+        draftCtx.defaultCandidates = [];
+        draftCtx.functionHint = formulaName;
+      });
     },
     [setContext]
   );