+++ packages/react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @fileverse-dev/fortune-react
 
+## 1.2.55
+
+### Patch Changes
+
+- Minor fix for tab on fx editor
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.55
+
+## 1.2.54
+
+### Patch Changes
+
+- NA
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.54
+
+## 1.2.53
+
+### Patch Changes
+
+- enable tab for header formula select
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.53
+
+## 1.2.52
+
+### Patch Changes
+
+- Scroll for mobile, comment edit box on hover, scroll enable for function container
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.2.52
+
 ## 1.2.51
 
 ### Patch Changes